Refactoring `webpack` config: multiple compilers, use `webpack-dev-server` (#15028)
* Using multiple compilers for webpack build.

* Use single compiler, feed plugin names into template.

* Adding back reload option.

* Using `webpack-dev-server`.

* Converting webpack configs to TS.

* Adding license header.

* Migrating `core` module back to JS`.

* Fixing up handling of `config.js`.

* Fixing linter hints.

* Removing unused import.

* Removing unused HMR plugin.

* Using `$GRAYLOG_API_URL` for proxying if present.

* Removing unused packages.

* Typing `global.pluginNames`.

* Removing `webpack-dev-middleware` & `webpack-hot-middleware`.

* Adding license header.

* Bumping `csstype` to `3.1.1`.
