+++ graylog2-web-interface/build-types.ts
@@ -14,13 +14,12 @@
  * along with this program. If not, see
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
-window.appConfig = {
-  gl2ServerUrl: '/api',
-  gl2AppPathPrefix: '',
-  rootTimeZone: 'UTC',
-  pluginUISettings: {
-    'org.graylog.plugins.customization.theme': {},
-    'org.graylog.plugins.customization.notifications': {},
-  },
-  isCloud: false,
-};
+declare global {
+  namespace NodeJS {
+    interface Global {
+      pluginNames: Array<string>
+    }
+  }
+}
+
+export {};
