+++ graylog2-web-interface/package.json
@@ -10,8 +10,8 @@
   },
   "readme": "../README.md",
   "scripts": {
-    "start": "node devServer.js",
-    "build": "cross-env disable_plugins=true webpack --config webpack.bundled.js",
+    "start": "webpack serve -c webpack.bundled.ts",
+    "build": "cross-env disable_plugins=true webpack --config webpack.bundled.ts",
     "build:stats": "cross-env disable_plugins=true webpack --profile --json > ./webpack/stats-$(git rev-parse --abbrev-ref HEAD).json",
     "build:analyze": "yarn build:stats && webpack-bundle-analyzer ./webpack/stats-$(git rev-parse --abbrev-ref HEAD).json ./target/web/build --mode=\"static\" --report=\"./webpack/analysis-$(git rev-parse --abbrev-ref HEAD).html\"",
     "generate:apidefs": "ts-node src/generator/generate.ts ../graylog2-server/target/swagger target/api",
@@ -138,8 +138,6 @@
     "@types/ua-parser-js": "^0.7.36",
     "@types/urijs": "^1.19.19",
     "assets-webpack-plugin": "^7.1.1",
-    "compression": "^1.7.2",
-    "connect-history-api-fallback": "^2.0.0",
     "copy-webpack-plugin": "^11.0.0",
     "cors": "^2.8.5",
     "cross-env": "^7.0.2",
@@ -151,7 +149,6 @@
     "estraverse-fb": "^1.3.1",
     "express": "^4.16.4",
     "express-formidable": "^1.2.0",
-    "express-http-proxy": "^1.6.2",
     "file-loader": "^6.0.0",
     "fork-ts-checker-webpack-plugin": "^8.0.0",
     "form-data": "^4.0.0",
@@ -169,10 +166,8 @@
     "utility-types": "^3.10.0",
     "webpack-bundle-analyzer": "^4.5.0",
     "webpack-cli": "5.0.1",
-    "webpack-dev-middleware": "^6.0.0",
-    "webpack-hot-middleware": "^2.25.1",
-    "webpack-merge": "^4.2.2",
-    "yargs": "^17.0.1"
+    "webpack-dev-server": "^4.13.1",
+    "webpack-merge": "^4.2.2"
   },
   "resolutions": {
     "immer": "9.0.6",
