+++ graylog2-web-interface/packages/graylog-web-plugin/src/PluginWebpackConfig.js
@@ -20,6 +20,8 @@ const webpack = require('webpack');
 const merge = require('webpack-merge');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 
+const TARGET = process.env.npm_lifecycle_event || 'build';
+
 function getPluginFullName(fqcn) {
   return `plugin.${fqcn}`;
 }
@@ -27,37 +29,40 @@ function getPluginFullName(fqcn) {
 function PluginWebpackConfig(defaultRootPath, fqcn, _options, additionalConfig) {
   const defaultOptions = {
     root_path: defaultRootPath,
+    src_path: path.resolve(defaultRootPath, 'src/web'),
     entry_path: path.resolve(defaultRootPath, 'src/web/index.jsx'),
     build_path: path.resolve(defaultRootPath, 'target/web/build'),
   };
 
   const options = merge(defaultOptions, _options);
-  // eslint-disable-next-line global-require,import/no-dynamic-require
-  const VENDOR_MANIFEST = require(path.resolve(_options.web_src_path, 'manifests', 'vendor-manifest.json'));
+  /* eslint-disable global-require,import/no-dynamic-require */
+  const VENDOR_MANIFEST = require(path.resolve(options.web_src_path, 'manifests', 'vendor-manifest.json'));
+  const core = require(path.resolve(options.web_src_path, 'webpack/core'));
+  const supportedBrowsers = require(path.resolve(options.web_src_path, 'supportedBrowsers'));
+  /* eslint-enable global-require,import/no-dynamic-require */
 
   const plugins = [
     new webpack.DllReferencePlugin({ manifest: VENDOR_MANIFEST, context: options.root_path }),
-    new HtmlWebpackPlugin({ filename: `${getPluginFullName(fqcn)}.module.json`, inject: false, template: path.resolve(_options.web_src_path, 'templates', 'module.json.template') }),
+    new HtmlWebpackPlugin({ filename: `${getPluginFullName(fqcn)}.module.json`, inject: false, template: path.resolve(options.web_src_path, 'templates', 'module.json.template') }),
   ];
+  const fullPluginName = getPluginFullName(fqcn);
 
-  const config = merge.smart(
-    // eslint-disable-next-line global-require,import/no-dynamic-require
-    require(path.resolve(_options.web_src_path, 'webpack.config.js')),
-    {
-      output: {
-        path: options.build_path,
-      },
-      plugins: plugins,
-      resolve: {
-        modules: [path.resolve(options.entry_path, '..')],
-      },
+  const config = merge.smart({
+    name: fullPluginName,
+    dependencies: ['vendor'],
+    entry: {
+      [fullPluginName]: options.entry_path,
     },
+    output: {
+      path: options.build_path,
+    },
+    plugins: plugins,
+    resolve: {
+      modules: [path.resolve(options.web_src_path, 'src'), path.resolve(options.web_src_path, 'node_modules')],
+    },
+  },
+  core.config(TARGET, options.src_path, options.root_path, options.web_src_path, supportedBrowsers),
   );
-
-  const entry = {};
-  entry[getPluginFullName(fqcn)] = options.entry_path;
-
-  config.entry = entry;
 
   if (additionalConfig) {
     return merge.smart(config, additionalConfig);
