+++ graylog2-web-interface/templates/index.html.template
@@ -8,6 +8,9 @@
     <script defer src="/<%= vendorModule().files.js[vendorChunk] %>"></script>
     <% } %>
     <%= htmlWebpackPlugin.tags.headTags %>
+    <% for (var idx in pluginNames()) { %>
+    <script defer src="/<%= pluginNames()[idx] %>.js"></script>
+    <% } %>
   </head>
   <body>
     <%= htmlWebpackPlugin.tags.bodyTags %>
