+++ graylog2-web-interface/webpack/vendor-module-ids.json
@@ -117,6 +117,7 @@
       "./node_modules/@emotion/unitless/dist/unitless.browser.esm.js": 351,
       "./node_modules/@emotion/utils/dist/emotion-utils.browser.esm.js": 444,
       "./node_modules/@emotion/weak-memoize/dist/weak-memoize.browser.esm.js": 220,
+      "./node_modules/@remix-run/router/dist/router.js": 599,
       "./node_modules/@tanstack/react-query/build/umd/index.production.js": 471,
       "./node_modules/chroma-js/chroma.js": 5792,
       "./node_modules/classnames/index.js": 4184,
@@ -528,12 +529,16 @@
       "./node_modules/react-redux/es/utils/warning.js": 8245,
       "./node_modules/react-redux/node_modules/react-is/cjs/react-is.production.min.js": 8359,
       "./node_modules/react-redux/node_modules/react-is/index.js": 2973,
+      "./node_modules/react-router-bootstrap/LinkContainer.js": 261,
+      "./node_modules/react-router-bootstrap/index.js": 860,
       "./node_modules/react-router-bootstrap/lib/IndexLinkContainer.js": 199,
       "./node_modules/react-router-bootstrap/lib/LinkContainer.js": 3610,
       "./node_modules/react-router-bootstrap/lib/index.js": 915,
+      "./node_modules/react-router-dom/dist/index.js": 655,
       "./node_modules/react-router-dom/esm/react-router-dom.js": 727,
       "./node_modules/react-router-dom/esm/react-router-dom.js|31d5e9d85c289ae7bc5f646483e4d91b": 9924,
       "./node_modules/react-router-dom/esm/react-router-dom.js|f87c77e6daa53d57f46e9108aab19cf8": 76,
+      "./node_modules/react-router/dist/index.js": 335,
       "./node_modules/react-router/esm/react-router.js": 6550,
       "./node_modules/react-router/esm/react-router.js|32a177570954fe8aee09cce6f8c45fd9": 20,
       "./node_modules/react-router/esm/react-router.js|59da0427b285e9b4bbe0ccc5e82a35cf": 977,
@@ -595,6 +600,9 @@
       "javascript/esm|./node_modules/@babel/runtime/helpers/esm/objectSpread2.js": 413,
       "javascript/esm|./node_modules/@babel/runtime/helpers/esm/objectWithoutPropertiesLoose.js": 3366,
       "javascript/esm|./node_modules/@babel/runtime/helpers/esm/setPrototypeOf.js": 9611,
+      "javascript/esm|./node_modules/@babel/runtime/helpers/esm/toPrimitive.js": 512,
+      "javascript/esm|./node_modules/@babel/runtime/helpers/esm/toPropertyKey.js": 343,
+      "javascript/esm|./node_modules/@babel/runtime/helpers/esm/typeof.js": 2,
       "javascript/esm|./node_modules/@tanstack/query-core/build/lib/focusManager.mjs": 5761,
       "javascript/esm|./node_modules/@tanstack/query-core/build/lib/hydration.mjs": 7224,
       "javascript/esm|./node_modules/@tanstack/query-core/build/lib/index.mjs": 9643,
@@ -2133,6 +2141,7 @@
       "node_modules/webpack/node_modules/timers-browserify/main.js": 611
     },
     "usedIds": [
+      2,
       5,
       8,
       19,
@@ -2154,7 +2163,6 @@
       176,
       180,
       188,
-      199,
       200,
       211,
       217,
@@ -2163,11 +2171,14 @@
       234,
       236,
       250,
+      261,
       277,
       302,
       316,
       329,
+      335,
       338,
+      343,
       345,
       351,
       354,
@@ -2187,6 +2198,7 @@
       490,
       504,
       511,
+      512,
       532,
       533,
       572,
@@ -2195,6 +2207,7 @@
       588,
       591,
       596,
+      599,
       602,
       609,
       610,
@@ -2202,7 +2215,7 @@
       622,
       626,
       644,
-      658,
+      655,
       667,
       670,
       682,
@@ -2210,11 +2223,11 @@
       696,
       699,
       702,
-      727,
       819,
       826,
       837,
       841,
+      860,
       867,
       874,
       877,
@@ -2222,7 +2235,6 @@
       894,
       906,
       913,
-      915,
       917,
       928,
       942,
@@ -2306,7 +2318,6 @@
       2138,
       2161,
       2175,
-      2177,
       2190,
       2193,
       2199,
@@ -2406,7 +2417,6 @@
       3528,
       3557,
       3607,
-      3610,
       3625,
       3651,
       3710,
@@ -2460,7 +2470,6 @@
       4502,
       4504,
       4518,
-      4523,
       4575,
       4578,
       4584,
@@ -2518,7 +2527,6 @@
       5380,
       5415,
       5427,
-      5429,
       5451,
       5466,
       5471,
@@ -2550,7 +2558,6 @@
       5805,
       5816,
       5822,
-      5826,
       5831,
       5834,
       5853,
@@ -2589,7 +2596,6 @@
       6524,
       6526,
       6539,
-      6550,
       6562,
       6567,
       6573,
@@ -2698,7 +2704,6 @@
       8236,
       8245,
       8254,
-      8273,
       8285,
       8286,
       8314,
@@ -2790,7 +2795,6 @@
       9681,
       9694,
       9728,
-      9731,
       9740,
       9755,
       9763,
